Added ProjectFileParsingManager to include git ignore parsing, project file collection.
